requirements.txt
django==4.1.1
djangorestframework==3.14.0
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos